requirements.txt
flask
Flask-SQLAlchemy
flask-login
pillow
gradio_client
bytesio
os